% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/isBIG5.R
\name{isBIG5}
\alias{isBIG5}
\title{Indicate whether the encoding of input string is BIG5.}
\usage{
isBIG5(string, combine = FALSE)
}
\arguments{
\item{string}{A character vector.}

\item{combine}{Whether to combine all the strings.}
}
\value{
Logical value.
}
\description{
Indicate whether the encoding of input string is BIG5.
}
\author{
Jian Li <\email{rweibo@sina.com}>
}

