% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/isGB18030.R
\name{isGB18030}
\alias{isGB18030}
\title{Indicate whether the encoding of input string is GB18030.}
\usage{
isGB18030(string, combine = FALSE)
}
\arguments{
\item{string}{A character vector.}

\item{combine}{Whether to combine all the strings.}
}
\value{
Logical value.
}
\description{
Indicate whether the encoding of input string is GB18030.
}
\author{
Jian Li <\email{rweibo@sina.com}>
}

