% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/isGB2312.R
\name{isGB2312}
\alias{isGB2312}
\title{Indicate whether the encoding of input string is GB2312.}
\usage{
isGB2312(string, combine = FALSE)
}
\arguments{
\item{string}{A character vector.}

\item{combine}{Whether to combine all the strings.}
}
\value{
Logical value.
}
\description{
Indicate whether the encoding of input string is GB2312.
}
\author{
Jian Li <\email{rweibo@sina.com}>
}

