% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/isGBK.R
\name{isGBK}
\alias{isGBK}
\title{Indicate whether the encoding of input string is GBK.}
\usage{
isGBK(string, combine = FALSE)
}
\arguments{
\item{string}{A character vector.}

\item{combine}{Whether to combine all the strings.}
}
\value{
Logical value.
}
\description{
Indicate whether the encoding of input string is GBK.
}
\author{
Jian Li <\email{rweibo@sina.com}>
}

