% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/isUTF8.R
\name{isUTF8}
\alias{isUTF8}
\title{Indicate whether the encoding of input string is UTF-8.}
\usage{
isUTF8(string, combine = FALSE)
}
\arguments{
\item{string}{A character vector.}

\item{combine}{Whether to combine all the strings.}
}
\value{
Logical value.
}
\description{
Indicate whether the encoding of input string is UTF-8.
}
\author{
Jian Li <\email{rweibo@sina.com}>
}

