% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/toUTF8.R
\name{toUTF8}
\alias{toUTF8}
\title{Convert encoding of Chinese string to UTF-8.}
\usage{
toUTF8(cnstring)
}
\arguments{
\item{cnstring}{A Chinese string vector.}
}
\value{
Converted vectors.
}
\description{
Convert encoding of Chinese string to UTF-8.
}
\author{
Jian Li <\email{rweibo@sina.com}>
}

